draft/requirements.txt
google-cloud-storage==1.43.0
pandas==1.3.4
requests==2.26.0
selenium==4.1.0
webdriver-manager==3.5.2


